+++ ArcaneWizardLibrary.lua
@@ -1,8 +1,8 @@
 local addonName, LIB = ...
 
 LIB.Internal = LIB.Internal or {}
+LIB.Internal.AddonChangelog = {}
 LIB.Internal.AddonLauncher = {}
-LIB.Internal.AddonUpdateNotice = {}
 
 ArcaneWizardLibrary = ArcaneWizardLibrary or {}